Rename crud service to sales-api and add improve docs
