+++ cmd/sidecar/tracer/handlers/zipkin.go
@@ -70,9 +70,9 @@ func (z *Zipkin) Publish(ctx context.Context, log *log.Logger, w http.ResponseWr
 	return nil
 }
 
-// send uses HTTP to send the data to the tracing sidecare for processing.
+// send uses HTTP to send the data to the tracing sidecar for processing.
 func (z *Zipkin) send(sendBatch []trace.SpanData) error {
-	le, err := newEndpoint("crud", z.localHost)
+	le, err := newEndpoint("sales-api", z.localHost)
 	if err != nil {
 		return err
 	}
