+++ dockerfile.sales-api
@@ -4,7 +4,7 @@ FROM golang:1.10.3 as build
 ENV CGO_ENABLED 0
 RUN mkdir -p /go/src/github.com/ardanlabs/service
 COPY . /go/src/github.com/ardanlabs/service
-WORKDIR /go/src/github.com/ardanlabs/service/cmd/crud
+WORKDIR /go/src/github.com/ardanlabs/service/cmd/sales-api
 RUN go build -ldflags "-s -w -X main.build=$(git rev-parse HEAD)" -a -tags netgo
 
 
@@ -15,12 +15,12 @@ ARG BUILD_DATE
 ARG VCS_REF
 EXPOSE 3000
 EXPOSE 4000
-COPY --from=build /go/src/github.com/ardanlabs/service/cmd/crud/crud /bin/
-ENTRYPOINT ["/bin/crud"]
+COPY --from=build /go/src/github.com/ardanlabs/service/cmd/sales-api/sales-api /bin/
+ENTRYPOINT ["/bin/sales-api"]
 
 LABEL org.opencontainers.image.created=$BUILD_DATE \
-      org.opencontainers.image.title="crud" \
+      org.opencontainers.image.title="sales-api" \
       org.opencontainers.image.authors="William Kennedy <bill@ardanlabs.com>" \
-      org.opencontainers.image.source="https://github.com/ardanlabs/service/cmd/crud" \
+      org.opencontainers.image.source="https://github.com/ardanlabs/service/cmd/sales-api" \
       org.opencontainers.image.revision=$VCS_REF \
       org.opencontainers.image.vendor="Ardan Labs"
